requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0
python-dotenv
gitpython
PyGithub
PyMuPDF
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
chromadb
sentence-transformers
google-generativeai
requests
python-multipart
